Printing last number #5

+++ 0x01-python-if_else_loops_functions/9-print_last_digit.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python3
 def print_last_digit(number):
-    str_n = str(number)[-1]
+    str_n = (int) str(number)[-1]
     print("{:d}".format(str_n), end="")
     return str_n
 # print("{}".format(print_last_digit(98)))
